modify separate source
- github_push.py from metadata_parser.py
